backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
weaviate-client>=4.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
websockets>=12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tiktoken>=0.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

